build: Move all VCS plugins to separate Gradle projects
This resolves #6602.
