+++ plugins/version-control-systems/git/src/main/resources/META-INF/services/org.ossreviewtoolkit.downloader.VersionControlSystem
@@ -1,0 +1,2 @@
+org.ossreviewtoolkit.plugins.versioncontrolsystems.git.Git
+org.ossreviewtoolkit.plugins.versioncontrolsystems.git.GitRepo
