+++ plugins/version-control-systems/subversion/src/main/resources/META-INF/services/org.ossreviewtoolkit.downloader.VersionControlSystem
@@ -1,0 +1,1 @@
+org.ossreviewtoolkit.plugins.versioncontrolsystems.subversion.Subversion
